split in memory repository adapter to separated configurations
