+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/Configuration.kt
@@ -1,6 +1,7 @@
 package github.com.hukuta94.delivery.api.startup.configuration
 
 import github.com.hukuta94.delivery.api.startup.configuration.adapter.KafkaConfiguration
+import github.com.hukuta94.delivery.api.startup.configuration.adapter.inmemory.InMemoryRepositoryConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.courier.CourierConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.order.OrderConfiguration
 import github.com.hukuta94.delivery.api.startup.configuration.scheduler.SchedulerConfiguration
@@ -20,5 +21,6 @@ import org.springframework.web.servlet.config.annotation.EnableWebMvc
     PortConfiguration::class,
     DomainEventHandlerConfiguration::class,
     KafkaConfiguration::class,
+    InMemoryRepositoryConfiguration::class,
 )
 open class Configuration
