+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/adapter/inmemory/InMemoryCourierRepositoryConfiguration.kt
@@ -1,4 +1,4 @@
-package github.com.hukuta94.delivery.api.startup.configuration.courier
+package github.com.hukuta94.delivery.api.startup.configuration.adapter.inmemory
 
 import github.com.hukuta94.delivery.core.application.event.DomainEventPublisher
 import github.com.hukuta94.delivery.infrastructure.adapter.inmemory.CourierInMemoryRepository
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-open class CourierRepositoryConfiguration {
+open class InMemoryCourierRepositoryConfiguration {
 
     @Bean
     open fun courierRepository(
