+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/adapter/inmemory/InMemoryOrderRepositoryConfiguration.kt
@@ -1,4 +1,4 @@
-package github.com.hukuta94.delivery.api.startup.configuration.order
+package github.com.hukuta94.delivery.api.startup.configuration.adapter.inmemory
 
 import github.com.hukuta94.delivery.core.application.event.DomainEventPublisher
 import github.com.hukuta94.delivery.infrastructure.adapter.inmemory.OrderInMemoryRepository
@@ -6,7 +6,7 @@ import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 
 @Configuration
-open class OrderRepositoryConfiguration {
+open class InMemoryOrderRepositoryConfiguration {
 
     @Bean
     open fun orderRepository(
