+++ api/src/main/kotlin/github/com/hukuta94/delivery/api/startup/configuration/courier/CourierConfiguration.kt
@@ -6,7 +6,6 @@ import org.springframework.context.annotation.Import
 @Configuration
 @Import(
     CourierUseCaseConfiguration::class,
-    CourierRepositoryConfiguration::class,
     CourierControllerConfiguration::class,
 )
 open class CourierConfiguration
